Add .travis.yml
Just build the library & the test, not the doc, as the current setup
requires a big TeX installation and I'm not sure yet if that's how
I want to do it.

Also fixed all uses of transpose, because I didn't understand how that
worked. Strange that my local gfortran didn't flag them.